requirements.txt
asgiref==3.2.10
Django==3.1.1
django-crispy-forms==1.9.2
django-safedelete==0.3.2
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
